[diffusion] refactor: collect component loaders to  component_loaders folder
