+++ python/sglang/multimodal_gen/runtime/loader/component_loaders/adapter_loader.py
